ddsim: move output level stuff to Helper.Output
